+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -98,6 +98,7 @@ export 'src/message_widget/components/stream_message_footer.dart';
 export 'src/message_widget/components/stream_message_header.dart';
 export 'src/message_widget/components/stream_message_leading.dart';
 export 'src/message_widget/stream_message_item.dart';
+export 'src/message_widget/stream_message_translation_store.dart';
 export 'src/message_widget/stream_moderated_message.dart';
 export 'src/message_widget/stream_quoted_message.dart';
 export 'src/message_widget/stream_system_message.dart';